+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/messenger/compose/view/room/settings/ExportRoom.android.kt
@@ -13,7 +13,6 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import de.connect2x.messenger.compose.view.DI
-import de.connect2x.messenger.compose.view.files.FilePathHelper
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
@@ -23,20 +22,20 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 import okio.Path.Companion.toPath
 import java.io.IOException
 
+
 private val log = KotlinLogging.logger { }
 
 @Composable
 internal actual fun SelectExportDestination(
     properties: FileBasedExportRoomProperties?,
-    result: (Destination?) -> Unit
+    result: (Destination?) -> Unit,
 ) {
     val appName = DI.get<MatrixMessengerConfiguration>().appName
     val i18n = DI.get<I18nView>()
     val context = LocalContext.current
 
-    val filePathHelper = remember { FilePathHelper(context) }
     val initialDirectory = remember {
-        filePathHelper.getAbsolutePath(
+        getAbsoluteDirectory(
             if (Build.VERSION.SDK_INT < 29) {
                 val permission = context.checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)
                 if (permission == PackageManager.PERMISSION_DENIED) {
@@ -45,10 +44,7 @@ internal actual fun SelectExportDestination(
                 Uri.fromFile(
                     Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).resolve(appName)
                 )
-            } else {
-                log.debug { "downloads folder: ${MediaStore.Downloads.EXTERNAL_CONTENT_URI}" }
-                MediaStore.Downloads.EXTERNAL_CONTENT_URI
-            }
+            } else MediaStore.Downloads.EXTERNAL_CONTENT_URI
         )
     }
 
@@ -60,3 +56,30 @@ internal actual fun SelectExportDestination(
     Text(i18n.exportRoomTargetDirectoryAndroid(), style = MaterialTheme.typography.bodyMedium)
 }
 
+private fun getAbsoluteDirectory(uri: Uri?): String {
+    val uriPath: String? = uri?.path
+    val default = Environment.DIRECTORY_DOWNLOADS
+    if (uriPath == null) log.warn { "invalid directory! defaulting to: $default" }
+    val directoryKey = uriPath?.split("/")?.also {
+        if (it.lastOrNull()?.contains('.') == true) log.warn { "directory uri might be a file: $uri" }
+    }?.getOrNull(1) ?: ""
+    val mappedDestination = directoryMap.entries.firstOrNull { entry ->
+        directoryKey.startsWith(entry.key)
+    }?.value ?: default
+    return "${Environment.getExternalStorageDirectory()}/${mappedDestination}"
+}
+
+private val directoryMap by lazy {
+    mapOf(
+        "music" to Environment.DIRECTORY_MUSIC,
+        "podcasts" to Environment.DIRECTORY_PODCASTS,
+        "ringtones" to Environment.DIRECTORY_RINGTONES,
+        "alarms" to Environment.DIRECTORY_ALARMS,
+        "notifications" to Environment.DIRECTORY_NOTIFICATIONS,
+        "pictures" to Environment.DIRECTORY_PICTURES,
+        "movies" to Environment.DIRECTORY_MOVIES,
+        "downloads" to Environment.DIRECTORY_DOWNLOADS,
+        "dcim" to Environment.DIRECTORY_DCIM,
+        "documents" to Environment.DIRECTORY_DOCUMENTS,
+    )
+}